fix: consolidate LDAIMetricSummary to single definition in model/types; rename tokens→usage
