+++ packages/sdk/server-ai/__tests__/LDAIConfigTrackerImpl.test.ts
@@ -621,7 +621,7 @@ it('summarizes tracked metrics', () => {
 
   expect(summary).toEqual({
     durationMs: 1000,
-    tokens: {
+    usage: {
       total: 100,
       input: 40,
       output: 60,
