+++ packages/sdk/server-ai/src/LDAIConfigTrackerImpl.ts
@@ -1,7 +1,7 @@
 import { LDContext } from '@launchdarkly/js-server-sdk-common';
 
 import { LDAIConfigTracker } from './api/config';
-import { LDAIMetricSummary } from './api/config/LDAIConfigTracker';
+import { LDAIMetricSummary } from './api/model/types';
 import { LDJudgeResult } from './api/judge/types';
 import {
   createBedrockTokenUsage,
@@ -333,13 +333,13 @@ export class LDAIConfigTrackerImpl implements LDAIConfigTracker {
   }
 
   trackTokens(tokens: LDTokenUsage): void {
-    if (this._trackedMetrics.tokens !== undefined) {
+    if (this._trackedMetrics.usage !== undefined) {
       this._ldClient.logger?.warn(
         'Token usage has already been tracked for this execution. Use createTracker() for a new execution.',
       );
       return;
     }
-    this._trackedMetrics.tokens = tokens;
+    this._trackedMetrics.usage = tokens;
     const trackData = this.getTrackData();
     if (tokens.total > 0) {
       this._ldClient.track('$ld:ai:tokens:total', this._context, trackData, tokens.total);
